prepared UI for adding images to game timeline
